ww/requirements.txt
appdirs==1.4.3
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
Django==1.11
django-filter==1.1.0
django-taggit==0.22.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-cloud-storage==1.1.1
google-resumable-media==0.0.2
googleapis-common-protos==1.5.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
oauth2client==4.1.2
packaging==16.8
protobuf==3.3.0
psycopg2==2.7.4
pyasn1==0.3.1
pyasn1-modules==0.0.10
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.3
pyparsing==2.2.0
pytz==2017.3
requests==2.18.1
rsa==3.4.2
six==1.10.0
urllib3==1.21.1
